% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_description.R
\docType{data}
\name{A.setosa}
\alias{A.setosa}
\title{Occurrence records of Alsophila setosa downloaded from Global Biodiversity
Information Facility (GBIF).}
\format{
A data frame with 508 rows and 45 variables
}
\source{
GBIF.org (08 July 2019) GBIF Occurrence Download \doi{10.15468/dl.6jesg0}
}
\usage{
A.setosa
}
\description{
A GBIF raw dataset containing 508 occurrence records for the tree fern
Alsophila setosa.
}
\keyword{datasets}
