% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_functions.R
\name{abrev.pttn}
\alias{abrev.pttn}
\title{Internal function of naturaList - Return abbreviation colapsed}
\usage{
abrev.pttn(df, line)
}
\arguments{
\item{df}{spec data frame provided in {\link[naturaList]{classify_occ}}}

\item{line}{specifies the line of the data frame to be collapsed}
}
\description{
Return colapsed abbreviation for a specific line of specialist data frame. It is
used as pattern in grep function inside {\link[naturaList]{classify_occ}}
}
\seealso{
{\link[naturaList]{classify_occ}}
}
