% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_description.R
\docType{data}
\name{BR}
\alias{BR}
\title{Brazil boundary}
\format{
A 'SpatialPolygonsDataFrame' with 1 feature
}
\usage{
BR
}
\description{
A spatial polygon with the Brazil boundaries
}
\keyword{datasets}
