% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_functions.R
\name{check.spec}
\alias{check.spec}
\title{Internal function of naturaList - Manual check of ambiguity in specialist
name}
\usage{
check.spec(class.occ, crit.levels, determined.by)
}
\arguments{
\item{class.occ}{internal data frame with observation classified according
{\link[naturaList]{classify_occ}} criteria}

\item{crit.levels}{crit.levels choosed by user in {\link[naturaList]{classify_occ}}}

\item{determined.by}{same as determined.by argumetn in {\link[naturaList]{classify_occ}}}
}
\description{
Creates interaction with user in which the user check if a string with the
identifier of a specimen has a specialist name. It solves ambiguity in classify
an occurence as identified by a specialist. It is used inside
{\link[naturaList]{classify_occ}}
}
\keyword{internal}
