% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify_occ.R
\name{classify_occ}
\alias{classify_occ}
\title{Classify occurrence records in levels of confidence in species determination}
\usage{
classify_occ(occ, spec = NULL, na.rm.coords = TRUE,
  crit.levels = c("det_by_spec", "taxonomist", "image", "sci_colection",
  "field_obs", "no_criteria_met"),
  institution.source = "institutionCode",
  collection.code = "collectionCode", catalog.number = "catalogNumber",
  year.event = "year", date.identified = "dateIdentified",
  scientific.name = "species", determined.by = "identifiedBy",
  longitude = "decimalLongitude", latitude = "decimalLatitude",
  basis.of.rec = "basisOfRecord", media.type = "mediaType",
  occ.id = "occurrenceID")
}
\arguments{
\item{occ}{dataframe with occurrence records information.}

\item{spec}{dataframe with specialists' names. See details.}

\item{na.rm.coords}{logical. If TRUE, remove occurrences with NA in latitude or longitude}

\item{crit.levels}{character. Vector with levels of confidence in decreasing order.
The criteria allowed are \code{det_by_spec}, \code{taxonomist},
\code{image}, \code{sci_colection}, \code{field_obs}, \code{no_criteria_met}. See datails.}

\item{institution.source}{column of \code{occ} with the name of the institution that provided the data}

\item{collection.code}{column of \code{occ} with the codes for institution names}

\item{catalog.number}{column of \code{occ} with catalog number}

\item{year.event}{column with of \code{occ} the year of the collection event}

\item{date.identified}{column of \code{occ} with the date of species determination}

\item{scientific.name}{column of \code{occ} with the species names}

\item{determined.by}{column of \code{occ} with the name of who determined the species}

\item{longitude}{column with of \code{occ} longitude in decimal degrees}

\item{latitude}{column with of \code{occ} latitude in decimal degrees}

\item{basis.of.rec}{column of \code{occ} with the recording types, as in GBIF. See details.}

\item{media.type}{column of \code{occ} with the media type of recording. See details.}

\item{occ.id}{column of \code{occ} with link or code for the occurence record.}
}
\description{
Classify occurrence records in levels of confidence in species determination
}
\details{
specialist dataframe must have one column for each name and abreviation.
First collumn must have the lastname. Subsequente colummns should have names and abbreviation
for each name. See an exemple with \code{data(specList)}

basis.of.rec is a character vector with one of the following types of record:
\code{PRESERVED_SPECIMEN} or \code{HUMAN_OBSERVATION}, as in GBIF data 'basisOfRecord'.

media.type uses the same pattern as GBIF data column mediaType indicating the existence
of an associated image by \code{STILLIMAGE}.
}
\author{
Arthur V. Rodrigues
}
