% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_eval.R
\name{clean_eval}
\alias{clean_eval}
\title{Evaluate the cleaning of occurrences records}
\usage{
clean_eval(
  occ.cl,
  geo.space,
  env.space,
  level.filter = c("1_det_by_spec"),
  r,
  scientific.name = "species",
  longitude = "decimalLongitude",
  latitude = "decimalLatitude"
)
}
\arguments{
\item{occ.cl}{Data frame with occurrence records information already
classified by \code{\link{classify_occ}} function.}

\item{geo.space}{a SpatialPolygons* or sf object defining the geographical space}

\item{env.space}{a SpatialPolygons* or sf object defining the environmental space.
Use the \code{\link{define_env_space}} for create this object.}

\item{level.filter}{a character vector including the levels in 'naturaList_levels'
collumn which filter the occurrence data set.}

\item{r}{a raster with 2 layers representing the environmental variables}

\item{scientific.name}{Column name of \code{occ.cl} with the species names.}

\item{longitude}{Column name of \code{occ.cl} longitude in decimal degrees.}

\item{latitude}{Column name of \code{occ.cl} latitude in decimal degrees.}
}
\value{
a list in which:

\code{area} data frame remaining area after cleaning proportional to the area
  before cleaning. The values vary from 0 to 1. Column named \code{r.geo.area}
  is the remaining area for all species in the geographic space and the
  \code{r.env.area} in the environmental space.

\code{comp} data frame with composition of species in sites (cells from raster
  layers) before cleaning (\code{comp$comp$BC}) and after cleaning
  (\code{comp$comp$AC}). The number of rows is equal the number of cells in
  \code{r}, and number of columns is equal to the number of species in the
  \code{occ.cl}.

\code{rich} data frame with a single column with the richness of each site

\code{site.coords} data frame with site's coordinates. It facilitates to built
  raster layers from results using \code{\link[raster]{rasterFromXYZ}}
}
\description{
This function compare the area occupied by
a species before and after pass through the cleaning procedure according to the chosen
level of filter.
The comparison can be made by measuring area in the geographical and in the environmental space
}
\examples{
\dontrun{
pkgs <- c("devtools", "raster", "maptools", "rnaturalearth", "sf")
lapply(pkgs, require, character.only = TRUE)

data("A.setosa")
data("speciaLists")

# classify
occ.cl <- classify_occ(A.setosa, speciaLists, spec.ambiguity = "not.spec")
## check points
occ.cl <-map_module(occ.cl) #delete points in the ocean

# download climate data
bioclim <- getData('worldclim', var='bio', res=10)

# Transform occurrence data in SpatialPointsDataFrame
spdf.occ.cl <- SpatialPoints(occ.cl[, c("decimalLongitude", "decimalLatitude")])

# redefine the extent of bioclim layers based on buffer around the occurrences
c.bioclim <- crop(bioclim, buffer(spdf.occ.cl, 100000)) # 100km buffer

# select two layers
raster.temp.prec <- c.bioclim[[c("bio1", "bio12")]]
df.temp.prec <- as.data.frame(raster.temp.prec)

### Define the environmental space for analysis
env.space <- define_env_space(df.temp.prec, buffer.size = 0.05)

# delimit the geographic space
# land area
land <- ne_countries(continent = 'south america')
un.land <- unionSpatialPolygons(land, land$scalerank)

# geo space based on crop
c.geo.space <- crop(un.land, c.bioclim)

#geo space based on intersect
i.geo.space <- intersect(un.land, buffer(spdf.occ.cl2, 200000))

### run the evaluation
cl.eval <- clean_eval(occ.cl,
                      env.space = env.space,
                      geo.space = c.geo.space,
                      r = raster.temp.prec)

#area results
cl.eval$area

### richness maps
## it makes sense if there are more than one species
rich.before.clean <- rasterFromXYZ(cbind(cl.eval$site.coords,
                                         cl.eval$rich$rich.BC))
rich.after.clean <- rasterFromXYZ(cbind(cl.eval$site.coords,
                                        cl.eval$rich$rich.AC))

plot(rich.before.clean)
plot(rich.after.clean)


### species area map
comp.bc <- as.data.frame(cl.eval$comp$comp.BC)
comp.ac <- as.data.frame(cl.eval$comp$comp.AC)

c.setosa.bc <- rasterFromXYZ(cbind(cbind(cl.eval$site.coords,
                                         comp.bc$`Cyathea setosa`)))
c.setosa.ac <- rasterFromXYZ(cbind(cbind(cl.eval$site.coords,
                                         comp.ac$`Cyathea setosa`)))

plot(c.setosa.bc)
plot(c.setosa.ac)

}

}
\seealso{
[define_env_space()]
}
