% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_description.R
\docType{data}
\name{cyathea.br}
\alias{cyathea.br}
\title{Occurrence records of Cyathea species in Brazil downloaded from Global Biodiversity
Information Facility (GBIF).}
\format{
A data frame with 3851 rows and 50 variables
}
\source{
GBIF.org (07 March 2021) GBIF Occurrence Download \doi{10.15468/dl.qrhynv}
}
\usage{
cyathea.br
}
\description{
A filtered GBIF dataset containing 3851 occurrence records for the fern species
from the genus Cyathea in Brazil. We filtered the data after download from GBIF
to ensure all occurrences records are from Brazil.
}
\keyword{datasets}
