% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define_env_space.R
\name{define_env_space}
\alias{define_env_space}
\title{Define environmental space}
\usage{
define_env_space(env, buffer.size, plot = TRUE)
}
\arguments{
\item{env}{matrix or data frame with two columns representing a bi-dimensional
environmental space.}

\item{buffer.size}{numeric value indicating a buffer size around each point
which will delimit the environmental space border. See details.}

\item{plot}{logical. whether to plot the polygon}
}
\value{
sfc_POLYGON
}
\description{
Based on two environmental variables, it defines a bi-dimensional environmental
space.
}
\details{
The environmental variables are standardized by range, which turns the
 range of each variable in 0 to 1. Then, it is delimited a buffer of size equal
 to\code{buffer.size} arround each point in this space. A polygon is returned
 indication the enviromental space based in the variables.
}
