% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_det_names.R
\name{get_det_names}
\alias{get_det_names}
\title{get the names in the 'determined.by' column}
\usage{
get_det_names(occ, determined.by = "identifiedBy")
}
\arguments{
\item{occ}{Data frame with occurrence records information.}

\item{determined.by}{Column name of \code{occ} with the name of who determined the
species.}
}
\description{
This function facilitates the search for non-taxonomist strings in the
'determined.by' column of occurrence records dataset
}
