% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid_filter.R
\name{grid_filter}
\alias{grid_filter}
\title{Filter the occurence with most confidence in species identification inside grid cells}
\usage{
grid_filter(occ.cl, grid.resolution = c(0.5, 0.5), r = NULL,
  institution.source = "institutionCode",
  collection.code = "collectionCode", catalog.number = "catalogNumber",
  year.event = "year", date.identified = "dateIdentified",
  scientific.name = "species", determined.by = "identifiedBy",
  longitude = "decimalLongitude", latitude = "decimalLatitude",
  basis.of.rec = "basisOfRecord", media.type = "mediaType",
  occ.id = "occurrenceID")
}
\arguments{
\item{occ.cl}{Data frame with occurrence records information already
classified by \code{\link{classify_occ}}.}

\item{grid.resolution}{Numeric vector with widht and height of grid cell in
decimal degrees.}

\item{r}{raster from which the grid cell resolution is derived.}

\item{institution.source}{Column name of \code{occ} with the name of the
institution that provided the data.}

\item{collection.code}{Column name of \code{occ} with the codes for institution
names.}

\item{catalog.number}{Column name of \code{occ} with catalog number.}

\item{year.event}{Column name of \code{occ} the year of the collection event.}

\item{date.identified}{Column name of \code{occ} with the date of species
determination.}

\item{scientific.name}{Column name of \code{occ} with the species names.}

\item{determined.by}{Column name of \code{occ} with the name of who determined the
species.}

\item{longitude}{Column name of \code{occ} longitude in decimal degrees.}

\item{latitude}{Column name of \code{occ} latitude in decimal degrees.}

\item{basis.of.rec}{Column name of \code{occ} with the recording types, as in GBIF.}

\item{media.type}{Column name of \code{occ} with the media type of recording.}

\item{occ.id}{Column name of \code{occ} with link or code for the occurence record.}
}
\value{
Data frame with the same columns of \code{occ.cl}.
}
\description{
Filter one occurence record per grid cell selected by the higher level of confidence
in species identification made by \code{\link{classify_occ}}
}
\examples{
occ.class <- classify_occ(A.setosa, speciaLists)
occ.grid <- grid.filter(occ.class)
occ.grid

}
\seealso{
\code{\link{classify_occ}}
}
\author{
Arthur V. Rodrigues
}
