% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_functions.R
\name{has.det.ID}
\alias{has.det.ID}
\title{Internal function of naturaList - Identifies if a occurence has a name for
the identifier of the specimen}
\usage{
has.det.ID(sp.df, ignore.det.names = NULL)
}
\arguments{
\item{sp.df}{reduced version of occurence data frame provided in
{\link[naturaList]{classify_occ}}}

\item{ignore.det.names}{ignore.det.names character vector indicatiing strings in the determined.by columns
that should be ignored as a taxonomist. See {\link[naturaList]{classify_occ}}.}
}
\description{
Identifies if a occurence has a name for the identifier of the specimen.
 It is used inside {\link[naturaList]{classify_occ}}
}
