% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_module.R
\name{map_module}
\alias{map_module}
\title{Check the occurrence records in a interactive map module}
\usage{
map_module(
  occ.cl,
  institution.source = "institutionCode",
  collection.code = "collectionCode",
  catalog.number = "catalogNumber",
  year.event = "year",
  date.identified = "dateIdentified",
  scientific.name = "species",
  determined.by = "identifiedBy",
  longitude = "decimalLongitude",
  latitude = "decimalLatitude",
  basis.of.rec = "basisOfRecord",
  media.type = "mediaType",
  occ.id = "occurrenceID"
)
}
\arguments{
\item{occ.cl}{Data frame with occurrence records information already
classified by \code{\link{classify_occ}} function.}

\item{institution.source}{Column name of \code{occ} with the name of the
institution that provided the data.}

\item{collection.code}{Column name of \code{occ} with the codes for institution
names.}

\item{catalog.number}{Column name of \code{occ} with catalog number.}

\item{year.event}{Column name of \code{occ} the year of the collection event.}

\item{date.identified}{Column name of \code{occ} with the date of species
determination.}

\item{scientific.name}{Column name of \code{occ} with the species names.}

\item{determined.by}{Column name of \code{occ} with the name of who determined the
species.}

\item{longitude}{Column name of \code{occ} longitude in decimal degrees.}

\item{latitude}{Column name of \code{occ} latitude in decimal degrees.}

\item{basis.of.rec}{Column name of \code{occ} with the recording types, as in GBIF.}

\item{media.type}{Column name of \code{occ} with the media type of recording.}

\item{occ.id}{Column name of \code{occ} with link or code for the occurence record.}
}
\value{
Data frame with the same columns of \code{occ.cl}.
}
\description{
Allows to delete occurrence records and to select occurence points by
classifcation levels or by drawing spatial polygons.
}
\examples{
occ.class <- classify_occ(A.setosa, speciaLists)
y
y
y
y
occ.selected <- map_module(occ.class)
occ.selected

}
\seealso{
\code{\link{classify_occ}}
}
\author{
Arthur V. Rodrigues
}
