% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_functions.R
\name{pttn.all.specialist}
\alias{pttn.all.specialist}
\title{Internal function of naturaList - Retturn specialists names in a collapesed
string}
\usage{
pttn.all.specialist(specialist)
}
\arguments{
\item{specialist}{specialist data frame}
}
\description{
Retturn specialists names in a collapesed string to be used in the internal
function {\link[naturaList]{specialist.conference}}
}
