% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_description.R
\docType{data}
\name{r.temp.prec}
\alias{r.temp.prec}
\title{Raster of temperature and precipitation}
\format{
A raster with two layers
}
\usage{
r.temp.prec
}
\description{
Raster of Annual Mean Temperature (bio1) and Total Annual Precipitation (bio2).
Layers were downloaded from worldclim database and cropped to the extent of
\code{cyathea_br} with a buffer of 100 km.
}
\keyword{datasets}
