% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_functions.R
\name{reduce.df}
\alias{reduce.df}
\title{Internal function of naturaList - reduce data.frama of occurence for a minimal
collumn length}
\usage{
reduce.df(
  df,
  institution.source = "institutionCode",
  collection.code = "collectionCode",
  catalog.number = "catalogNumber",
  year.event = "year",
  date.identified = "dateIdentified",
  scientific.name = "species",
  determined.by = "identifiedBy",
  longitude = "decimalLongitude",
  latitude = "decimalLatitude",
  basis.of.rec = "basisOfRecord",
  media.type = "mediaType",
  occ.id = "occurrenceID",
  na.rm.coords = TRUE
)
}
\arguments{
\item{df}{occurence data frame provided in {\link[naturaList]{classify_occ}}}

\item{institution.source}{institution.source = "institutionCode"}

\item{collection.code}{collection.code = "collectionCode"}

\item{catalog.number}{catalog.number = "catalogNumber"}

\item{year.event}{year.event = "year",}

\item{date.identified}{date.identified = "dateIdentified"}

\item{scientific.name}{scientific.name = "species"}

\item{determined.by}{determined.by = "identifiedBy"}

\item{longitude}{longitude = "decimalLongitude"}

\item{latitude}{latitude = "decimalLatitude"}

\item{basis.of.rec}{basis.of.rec = "basisOfRecord"}

\item{media.type}{media.type = "mediaType"}

\item{occ.id}{occ.id = "occurrenceID"}

\item{na.rm.coords}{= TRUE}
}
\description{
Reduce columns of occurrence data.frame required by
{\link[naturaList]{classify_occ}} to facilitate internal operation
}
\seealso{
{\link[naturaList]{classify_occ}}
}
