% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_functions.R
\name{rm.coord.dup}
\alias{rm.coord.dup}
\title{Internal funciton of naturaList - Remove duplicate occurrence}
\usage{
rm.coord.dup(x, latitude, longitude)
}
\arguments{
\item{x}{data frame with filtered occurrences}

\item{latitude}{name of column with latitude}

\item{longitude}{name of column with longitude}
}
\description{
Remove duplicated occurrence based on coordinates. It is used in
{\link[naturaList]{grid_filter}}
}
