% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_description.R
\docType{data}
\name{spec_names_ex}
\alias{spec_names_ex}
\title{Example of specialist names with accent marks}
\format{
character
}
\usage{
spec_names_ex
}
\description{
Example of specialist names with accent marks
}
\keyword{datasets}
