% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_functions.R
\name{specialist.conference}
\alias{specialist.conference}
\title{Internal function of naturaList - Confirm if an occurrence record was identified by
a specialist without ambiguity}
\usage{
specialist.conference(pt.df, specialist)
}
\arguments{
\item{pt.df}{a line of the reduced version of the occurrence data frame}

\item{specialist}{specialist data frame}
}
\description{
Confirm if an occurrence record was identified by
a specialist without ambiguity. It is used inside
 {\link[naturaList]{classify_occ}}
}
\keyword{internal}
