% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_description.R
\docType{data}
\name{speciaLists}
\alias{speciaLists}
\title{Specialists of ferns and lycophytes of Brazil}
\format{A data frame with 27 rows and 8 columns}
\usage{
speciaLists
}
\description{
A dataset containing the specialists of ferns and lycophytes of Brazil formated
to be used by \code{naturaList} package
}
\keyword{datasets}
