% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_functions.R
\name{verify.specialist}
\alias{verify.specialist}
\title{Internal function of naturaList - Verify if a string has unambiguous specialist
name}
\usage{
verify.specialist(pattern, string)
}
\arguments{
\item{pattern}{a pattern from {\link[naturaList]{pttn.all.specialist}} function}

\item{string}{string with the name of who identified the specimen}
}
\value{
character. \code{""} or \code{"_verify"}.
}
\description{
Based on pattern generated by {\link[naturaList]{pttn.all.specialist}} it
verifies if a string has unambiguous specialist name. It is used in internal
function {\link[naturaList]{specialist.conference}}
}
\keyword{internal}
