% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_functions.R
\name{abrev.pttn}
\alias{abrev.pttn}
\title{Internal function of naturaList - Return abbreviation collapsed}
\usage{
abrev.pttn(df, line)
}
\arguments{
\item{df}{spec data frame provided in {\link[naturaList]{classify_occ}}}

\item{line}{specifies the line of the data frame to be collapsed}
}
\value{
a list with two elements in regex format:
  \item{[[1]]}{the abbreviation of the first name;}
  \item{[[2]]}{regex pattern with all names and abbreviations.}
}
\description{
Return collapsed abbreviation for a specific line of specialist data frame.
It is used as pattern in grep function inside
{\link[naturaList]{classify_occ}}
}
\seealso{
{\link[naturaList]{classify_occ}}
}
\keyword{internal}
