% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify_occ.R
\name{classify_occ}
\alias{classify_occ}
\title{Classify occurrence records in levels of confidence in species determination}
\usage{
classify_occ(occ, spec = NULL, na.rm.coords = TRUE,
  crit.levels = c("det_by_spec", "taxonomist", "image", "sci_colection",
  "field_obs", "no_criteria_met"),
  institution.source = "institutionCode",
  collection.code = "collectionCode", catalog.number = "catalogNumber",
  year.event = "year", date.identified = "dateIdentified",
  scientific.name = "species", determined.by = "identifiedBy",
  longitude = "decimalLongitude", latitude = "decimalLatitude",
  basis.of.rec = "basisOfRecord", media.type = "mediaType",
  occ.id = "occurrenceID")
}
\arguments{
\item{occ}{Data frame with occurrence records information.}

\item{spec}{Data frame with specialists' names. See details.}

\item{na.rm.coords}{Logical. If TRUE, remove occurrences with NA in latitude
or longitude}

\item{crit.levels}{Character. Vector with levels of confidence in decreasing
order. The criteria allowed are \code{det_by_spec}, \code{taxonomist},
\code{image}, \code{sci_colection}, \code{field_obs}, \code{no_criteria_met}.
 See datails.}

\item{institution.source}{Column name of \code{occ} with the name of the
institution that provided the data.}

\item{collection.code}{Column name of \code{occ} with the codes for institution
names.}

\item{catalog.number}{Column name of \code{occ} with catalog number.}

\item{year.event}{Column name of \code{occ} the year of the collection event.}

\item{date.identified}{Column name of \code{occ} with the date of species
determination.}

\item{scientific.name}{Column name of \code{occ} with the species names.}

\item{determined.by}{Column name of \code{occ} with the name of who determined the
species.}

\item{longitude}{Column name of \code{occ} longitude in decimal degrees.}

\item{latitude}{Column name of \code{occ} latitude in decimal degrees.}

\item{basis.of.rec}{Column name of \code{occ} with the recording types, as in GBIF.
See details.}

\item{media.type}{Column name of \code{occ} with the media type of recording.
See details.}

\item{occ.id}{Column name of \code{occ} with link or code for the occurence record.}
}
\value{
The \code{occ} data frame plus the classification of each record
in a new column, named \code{naturaList_levels}.
}
\description{
Classify occurrence records in levels of confidence in species
determination
}
\details{
\code{spec} data frame must have columns separating \code{LastName},
\code{Name} and \code{Abbrev}. The first column contain the \code{LastName} of the specialists. The
following columns should have names (\code{Name}) and abbreviation
(\code{Abbrev}) for each specialist. See an exemple in \code{\link[naturaList]{speciaList}} data.

\code{basis.of.rec} is a character vector with one of the following
types of record:\code{PRESERVED_SPECIMEN} or \code{HUMAN_OBSERVATION}, as in
GBIF data 'basisOfRecord'.

\code{media.type} uses the same pattern as GBIF mediaType column,
indicating the existence of an associated image with \code{STILLIMAGE}.
}
\examples{
occ.class <- classify_occ(A.setosa, speciaLists)
y
y
y
y
occ.class
}
\seealso{
\code{\link[naturaList]{speciaList}}
}
\author{
Arthur V. Rodrigues
}
