% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_spec_df.R
\name{create_spec_df}
\alias{create_spec_df}
\title{Create specialist data frame from character vector}
\usage{
create_spec_df(spec.char)
}
\arguments{
\item{spec.char}{A character vector of specialist names}
}
\description{
Create a specialist data frame for ready for use in naturaList::classify_occ()
from a character vector containg the specialists names
}
\examples{
spec_list <- c("Caetano Veloso", "Antônio Carlos Tom Jobim",
"Gilberto Gil", "Vinícius de Morais")

create_spec_df(spec_list)

}
