% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_functions.R
\name{make.polygon}
\alias{make.polygon}
\title{Internal function of naturaList - Create SpatialPolygons from a list of coordinates}
\usage{
make.polygon(df)
}
\arguments{
\item{df}{a data frame provided by {\link{pol.coords}}}
}
\value{
a \code{SpatialPolygon} object
}
\description{
Create SpatialPolygons from a list of coordinates. It is used in {\link{map_module}}
}
\keyword{internal}
