% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_functions.R
\name{pttn.all.specialist}
\alias{pttn.all.specialist}
\title{Internal function of naturaList - Return specialists names in a collapsed
string}
\usage{
pttn.all.specialist(specialist)
}
\arguments{
\item{specialist}{specialist data frame}
}
\value{
character. A regex pattern for the specialist full name
}
\description{
Return specialists names in a collapsed string to be used in the internal
function {\link[naturaList]{specialist.conference}}
}
\keyword{internal}
