% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_functions.R
\name{rm.coord.dup}
\alias{rm.coord.dup}
\title{Internal function of naturaList - Remove duplicate occurrence}
\usage{
rm.coord.dup(x, decimal.latitude, decimal.longitude)
}
\arguments{
\item{x}{data frame with filtered occurrences}

\item{decimal.latitude}{name of column with decimal.latitude}

\item{decimal.longitude}{name of column with decimal.longitude}
}
\value{
data frame with occurrence records
}
\description{
Remove duplicated occurrence based on coordinates. It is used in
{\link[naturaList]{grid_filter}}
}
\keyword{internal}
